backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
pytesseract==0.3.10; sys_platform == "win32"
nltk==3.8.1